Debugged summary page form
