+++ package.json
@@ -39,12 +39,14 @@
     "gapi-script": "^1.2.0",
     "lodash": "^4.17.21",
     "lodash.isempty": "^4.4.0",
+    "moment": "^2.30.1",
     "nativewind": "^4.1.23",
     "prop-types": "^15.8.1",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-native": "^0.79.4",
     "react-native-calendars": "^1.1312.1",
+    "react-native-date-picker": "^5.0.13",
     "react-native-gesture-handler": "~2.24.0",
     "react-native-paper": "^5.14.5",
     "react-native-reanimated": "~3.17.4",
