+++ src/app/(tabs)/itinerary.tsx
@@ -1,163 +1,464 @@
+import AsyncStorage from '@react-native-async-storage/async-storage';
 import DateTimePicker from '@react-native-community/datetimepicker';
-import React, { useState } from 'react';
-import { Button, Pressable, ScrollView, StatusBar, StyleSheet, Text, TextInput, View } from 'react-native';
-import { SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
+import { useFocusEffect } from '@react-navigation/native';
+import { addDoc, collection, deleteDoc, doc, getDocs, Timestamp, updateDoc } from 'firebase/firestore';
+import React, { useMemo, useRef, useState } from 'react';
+import { Alert, Button, Modal, Platform, StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
+import { AgendaList, CalendarProvider, ExpandableCalendar, WeekCalendar } from 'react-native-calendars';
+import { SafeAreaView } from 'react-native-safe-area-context';
+import { auth, db } from '../../../firebaseConfig';
+import AgendaItem from '../mocks/AgendaItem';
+import { getTheme, lightThemeColor, themeColor } from '../mocks/theme';
+import testIDs from '../testIDs';
 
+const leftArrowIcon = require('../img/previous.png');
+const rightArrowIcon = require('../img/next.png');
 
-//type annotations
-type Activity = { //(tuples)
-  time: string;
-  activity: string;
-};
-
-type Day = {
-  id: number; //day number of the trip
-  activities: Activity[]; //array of tuples
-};
-
-export default function ItineraryScreen() {
-  const [itinerary, setItinerary] = useState([
-    { id: 1, activities: [{ time: '', activity: '' }] } //initalise day 1 with 1 empty time slot (array of tuples: {time, activity})
-  ]);
-  const [showPicker, setShowPicker] = useState<{  //tracks which time slot of which day is currrently being edited
-    dayId: number | null;
-    index: number | null;
-  }>({ dayId: null, index: null });
-
-  const [selectedTime, setSelectedTime] = useState<Date>(new Date()); //to store the most recently selected time
-  
-  const onTimeChange = (event: any, date?: Date) => { //when a user selects a time
-    if (!date) return;
-    const timeString = date.toTimeString().slice(0, 5); //time formatting
-    setItinerary(prev =>
-      prev.map(day =>
-        day.id === showPicker.dayId
-          ? {
-              ...day,
-              slots: day.activities.map((slot: Activity, i: number) =>
-                i === showPicker.index ? { ...slot, time: timeString } : slot
-              ),
-            }
-          : day
-      )
-    ); //updates the time of the specific slot with matching dayId and index
-    setShowPicker({ dayId: null, index: null }); // hide time picker
-  };
-
-  const addDay = () => { //adds a new day with a single blank time slot
-    setItinerary([...itinerary, { id: Date.now(), activities: [] }]);
-  };
-
-  const removeDay = (id: number) => { //removes a specific day with the matching dayId
-    setItinerary(itinerary.filter(day => day.id !== id)); //only keep the days that do not match the specified dayId
-  };
-
-  const addTimeSlot = (dayId: number) => {
-    setItinerary(itinerary.map(day =>
-      day.id === dayId
-        ? { ...day, activities: [...day.activities, { time: '', activity: '' }] } //to the current day. add another element to the array of activity tuples
-        : day //do nothing if there is already a slot at the specified dayId and index
-    ));
-  };
-
-  const updateActivity = (dayId: number, index: number, field: 'activity' | 'time', value: string) => {
-    setItinerary(itinerary.map(day =>
-      day.id === dayId
-        ? {
-            ...day,
-            activities: day.activities.map((slot, i) =>
-              i === index ? { ...slot, [field]: value } : slot //change the activity element of the tuple matching the dayId and index
-            )
-          }
-        : day
-    ));
-  };
-
-  return (
-    <SafeAreaProvider>
-      <SafeAreaView style={styles.containerScroll} edges={['top']}>
-    <ScrollView className='flex-1 py-12 px-7'>
-      {itinerary.map((day, dayIdx) => (
-        <View key={day.id} style={styles.dayContainer}>
-          <Text style={styles.dayTitle}>Day {dayIdx + 1}</Text>
-          {day.activities.map((slot, i) => (
-            <View key={i} style={styles.row}>
-              <Pressable onPress={() => setShowPicker({ dayId: day.id, index: i })}>
-                <Text style={styles.timeBox}>{slot.time || 'Set Time'}</Text>
-              </Pressable>
-              <TextInput //input the activity
-                style={styles.input}
-                placeholder="Activity"
-                value={slot.activity}
-                onChangeText={text => updateActivity(day.id, i, 'activity', text)}
-              />
-            </View>
-          ))}
-          <Button title="Add Time Slot" onPress={() => addTimeSlot(day.id)} />
-          <Button title="Delete Day" color="red" onPress={() => removeDay(day.id)} />
-        </View>
-      ))}
-      <Button title="Add Day" onPress={addDay} />
-
-      {showPicker.dayId !== null && (
-        <DateTimePicker
-          mode="time"
-          is24Hour={true}
-          value={new Date()}
-          onChange={(_, selectedTime) => {
-            if (selectedTime) {
-              updateActivity(showPicker.dayId!, showPicker.index!, 'time', selectedTime.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit', hour12: false }));
-            }
-            setShowPicker({ dayId: null, index: null });
-          }}
-        />
-      )}
-    </ScrollView>
-    </SafeAreaView>
-    </SafeAreaProvider>
-  );
+interface Props {
+  weekView?: boolean;
 }
 
+interface EventData {
+  id: string;
+  date: string;
+  activity: string;
+  hour: string;
+  duration: string;
+  address: string;
+}
+
+const AgendaInfiniteListScreen = (props: Props) => {
+  const [items, setItems] = useState<any[]>([]);
+  const [markedDates, setMarkedDates] = useState<Record<string, any>>({});
+  const [fVisible, setFVisible] = useState(false);
+
+  const [showDatePicker, setShowDatePicker] = useState(false);
+  const [showTimePicker, setShowTimePicker] = useState(false);
+
+  const [selectedEvent, setSelectedEvent] = useState<any | null>(null);
+  const [currentlyEditingEventId, setCurrentlyEditingEventId] = useState<string | null>(null);
+
+  const fetchEvents = async () => {
+    const tripId = await AsyncStorage.getItem('activeTripId');
+    if (!tripId) {
+      Alert.alert("Error", "Trip ID not found.");
+      return;
+    }
+
+    const user = auth.currentUser;
+    if (!user) return;
+    
+    try {
+      const snapshot = await getDocs(collection(db, 'trips', tripId, 'events'));
+      const docs: EventData[] = snapshot.docs.map(doc => ({
+        id: doc.id,
+        ...(doc.data() as Omit<EventData, 'id'>)
+      }));
+      const grouped: Record<string, any[]> = {};
+      docs.forEach(event => {
+        if (!event.date) return;
+        if (!grouped[event.date]) grouped[event.date] = [];
+
+        grouped[event.date].push({
+          title: event.activity,
+          hour: event.hour,
+          duration: event.duration,
+          address: event.address,
+          id: event.id,
+        });
+
+        Object.values(grouped).forEach(eventsArray => {
+          eventsArray.sort((a, b) => {
+          const timeA = a.hour?.split(':').map(Number) || [0, 0];
+          const timeB = b.hour?.split(':').map(Number) || [0, 0];
+          return timeA[0] - timeB[0] || timeA[1] - timeB[1];
+          });
+        });
+      });
+
+      const structured = Object.entries(grouped)
+        .sort(([a], [b]) => a.localeCompare(b))
+        .map(([date, data]) => {
+          const sortedData = [...data].sort((a, b) => {
+            const toMinutes = (time: string) => {
+              const match = time.match(/(\d+):(\d+)?(am|pm)/);
+              if (!match) return 0;
+              let [ , hourStr, minuteStr = '0', period ] = match;
+              let hour = parseInt(hourStr, 10);
+              const minute = parseInt(minuteStr, 10);
+              if (period === 'pm' && hour !== 12) hour += 12;
+              if (period === 'am' && hour === 12) hour = 0;
+              return hour * 60 + minute;
+            };
+            return toMinutes(a.hour) - toMinutes(b.hour);
+          });
+          return {
+            title: date,
+            data: sortedData,
+          };
+        });
+
+      setItems(structured);
+      const newMarked: Record<string, any> = {};
+      structured.forEach(item => {
+        if (item.data.length > 0) {
+          newMarked[item.title] = { marked: true };
+        }
+      });
+      setMarkedDates(newMarked);
+    } catch (err) {
+      console.error('Error loading events:', err);
+    }
+  };
+
+  useFocusEffect(
+    React.useCallback(() => {
+      if (auth.currentUser) {
+        fetchEvents();
+      }
+    }, [])
+  );
+
+  const [form, setForm] = useState({
+    date: new Date().toISOString().split('T')[0],
+    hour: '',
+    durationHours: '',
+    durationMinutes: '',
+    activity: '',
+    address: ''
+  });
+
+
+  const handleChange = (field: string, value: string) => {
+    setForm({ ...form, [field]: value });
+  };
+  const {weekView} = props;
+  const theme = useRef(getTheme());
+  const todayBtnTheme = useRef({
+    todayButtonTextColor: themeColor
+  });
+
+  const renderItem = ({ item }: any) => {
+    return (
+      <TouchableOpacity onPress={() => setSelectedEvent(item)}>
+        <AgendaItem item={item} />
+      </TouchableOpacity>
+    );
+  };
+
+  const addEvent = async () => {
+    const tripId = await AsyncStorage.getItem('activeTripId');
+    if (!tripId) {
+      Alert.alert("Error", "Trip ID not found.");
+      return;
+    }
+
+    const user = auth.currentUser;
+    if(!user) {
+      console.warn("user not signed in");
+      return;
+    }
+    try {
+      const today = new Date().toISOString().split('T')[0];
+      const newEvent = {
+        date: today,
+        hour: '12pm',
+        duration: '1h',
+        title: 'New Trip Activity',
+        createdBy: user.uid,
+        itemCustomHeightType: 'LongEvent',
+      };
+      await addDoc(collection(db, 'trips', tripId, 'events'), newEvent);
+      console.log('event added');
+  } catch (error) {
+    console.error('failed to add event:', error);
+  }
+};
+
+const handleSubmit = async () => {
+  const tripId = await AsyncStorage.getItem('activeTripId');
+  if (!tripId) {
+    Alert.alert("Error", "Trip ID not found.");
+    return;
+  }
+
+  const user = auth.currentUser;
+  if (!user) {
+    Alert.alert("Error", "User not signed in");
+    return;
+  }
+
+  try {
+    const isEditing = !!currentlyEditingEventId;
+
+    const eventData = {
+      date: form.date,
+      hour: form.hour,
+      duration: `${form.durationHours || 0}h ${form.durationMinutes || 0}m`,
+      address: form.address,
+      activity: form.activity,
+      createdBy: user.uid,
+      timestamp: Timestamp.now()
+    };
+
+    if (isEditing) {
+      const tripId = await AsyncStorage.getItem('activeTripId');
+      if (!tripId) return;
+        const eventRef = doc(db, 'trips', tripId, 'events', currentlyEditingEventId!);
+        await updateDoc(eventRef, eventData);
+    } else {
+        await addDoc(collection(db, 'trips', tripId, 'events'), eventData);
+      }
+    await fetchEvents();
+    Alert.alert("Success", isEditing ? "Event updated!" : "Event saved!");
+
+    setFVisible(false);
+    setSelectedEvent(null);
+    setForm({
+      date: new Date().toISOString().split('T')[0],
+      hour: '',
+      durationHours: '',
+      durationMinutes: '',
+      activity: '',
+      address: ''
+    });
+    setCurrentlyEditingEventId(null);
+
+  } catch (err) {
+    console.error('Error saving event:', err);
+    Alert.alert("Error", "Failed to save event.");
+  }
+};
+
+
+const marked = useMemo(() => {
+  const result: Record<string, any> = {};
+  items.forEach(item => {
+    if (item.data && item.data.length > 0) {
+      result[item.title] = { marked: true };
+    }
+  });
+  return result;
+}, [items]);
+
+
+  return (
+    <SafeAreaView style = {{ flex: 1 }}>
+      <View className='h-full justify-center py-1'>
+        <CalendarProvider
+          date={items[0]?.title || new Date().toISOString().split('T')[0]}
+          showTodayButton
+          theme={todayBtnTheme.current}>
+        {weekView ? (
+          <WeekCalendar testID={testIDs.weekCalendar.CONTAINER} firstDay={1} markedDates={markedDates}/>
+        ) : (
+          <ExpandableCalendar
+            testID={testIDs.expandableCalendar.CONTAINER}
+            firstDay={1}
+            markedDates={markedDates}
+            leftArrowImageSource={leftArrowIcon}
+            rightArrowImageSource={rightArrowIcon}/>
+        )}
+        <AgendaList
+          theme={theme.current}
+          sections={items}
+          renderItem={renderItem}
+          keyExtractor={(item) => item.id}
+          sectionStyle={styles.section}
+          infiniteListProps={{
+            itemHeight: 80,
+            titleHeight: 50,
+            itemHeightByType: {
+              LongEvent: 120
+            }
+          }}/>
+        <Modal visible={fVisible} animationType="slide" transparent={true}>
+        <View style={styles.modalContainer}>
+          <View style={styles.modalContent}>
+            <Text style={styles.modalTitle}>Add New Event</Text>
+            <Text style={styles.label}>Date</Text>
+            <TouchableOpacity onPress={() => setShowDatePicker(true)} style={styles.pickerInput}>
+              <Text>{form.date}</Text>
+            </TouchableOpacity>
+            {showDatePicker && (
+              <DateTimePicker
+                value={new Date(form.date)}
+                mode="date"
+                display="default"
+                onChange={(event, selectedDate) => {
+                  setShowDatePicker(false);
+                  if (selectedDate) {
+                    setForm({ ...form, date: selectedDate.toISOString().split('T')[0] });
+                  }
+                }}/>
+            )}
+            <Text style={styles.label}>Start Time</Text>
+            <TouchableOpacity onPress={() => setShowTimePicker(true)} style={styles.pickerInput}>
+              <Text>{form.hour || 'Select time'}</Text>
+            </TouchableOpacity>
+            {showTimePicker && (
+              <DateTimePicker
+                value={new Date()}
+                mode="time"
+                is24Hour={true}
+                display={Platform.OS === 'ios' ? 'spinner' : 'default'}
+                onChange={(event, selectedTime) => {
+                  if (Platform.OS === 'android') {
+                    setShowTimePicker(false);
+                  }
+                  if (selectedTime) {
+                    const hour = selectedTime.getHours();
+                    const minute = selectedTime.getMinutes();
+                    const formatted = `${hour.toString().padStart(2, '0')}:${minute
+                      .toString()
+                      .padStart(2, '0')}`;
+                    setForm({ ...form, hour: formatted });
+                    if (Platform.OS === 'ios') {
+                      setShowTimePicker(false);
+                    }
+                  }
+                }}/>
+            )}
+            <Text style={styles.label}>Duration</Text>
+            <View style={{ flexDirection: 'row', gap: 10 }}>
+              <TextInput
+                placeholder="Hrs"
+                keyboardType="numeric"
+                style={[styles.input, { width: 60 }]}
+                value={form.durationHours}
+                onChangeText={t => setForm({ ...form, durationHours: t })}/>
+              <TextInput
+                placeholder="Min"
+                keyboardType="numeric"
+                style={[styles.input, { width: 60 }]}
+                value={form.durationMinutes}
+                onChangeText={t => setForm({ ...form, durationMinutes: t })}/>
+            </View>
+            <TextInput placeholder="Activity Name" style={styles.input} value={form.activity} onChangeText={t => handleChange('activity', t)} />
+            <TextInput placeholder="Location" style={styles.input} value={form.address} onChangeText={t => handleChange('address', t)} />
+            <View style={{ flexDirection: 'row', justifyContent: 'space-between' }}>
+              <Button title="Cancel" onPress={() => setFVisible(false)} />
+              <Button title="Save" onPress={handleSubmit} />
+            </View>
+          </View>
+        </View>
+        </Modal>
+
+        <Modal visible={!!selectedEvent} animationType="slide" transparent={true}>
+        <View style={styles.modalContainer}>
+          <View style={styles.modalContent}>
+            <Text style={styles.modalTitle}>Event Details</Text>
+            <Text style={styles.label}>Activity: {selectedEvent?.title}</Text>
+            <Text style={styles.label}>Time: {selectedEvent?.hour}</Text>
+            <Text style={styles.label}>Duration: {selectedEvent?.duration}</Text>
+            <Text style={styles.label}>Address: {selectedEvent?.address}</Text>
+            <View style={{ flexDirection: 'row', justifyContent: 'space-between', marginTop: 20 }}>
+              <Button
+                title="Edit"
+                onPress={() => {
+                  setForm({
+                    date: items.find(i => i.data.includes(selectedEvent))?.title || new Date().toISOString().split('T')[0],
+                    hour: selectedEvent?.hour,
+                    durationHours: selectedEvent?.duration?.split('h')[0]?.trim() || '',
+                    durationMinutes: selectedEvent?.duration?.split('h')[1]?.replace('m', '')?.trim() || '',
+                    activity: selectedEvent?.title,
+                    address: selectedEvent?.address,
+                  });
+                  setCurrentlyEditingEventId(selectedEvent?.id)
+                  setFVisible(true);
+                  setSelectedEvent(null);
+                }}/>
+              <Button
+                title="Delete"
+                color="red"
+                onPress={async () => {
+                  const tripId = await AsyncStorage.getItem('activeTripId');
+                  if (!tripId) {
+                    Alert.alert("Error", "Trip ID not found.");
+                    return;
+                  }
+                  try {
+                    const user = auth.currentUser;
+                    if (!user || !selectedEvent?.id) return;
+
+                    const docRef = doc(db, 'trips', tripId, 'events', selectedEvent.id);
+                    await deleteDoc(docRef);
+
+                    Alert.alert("Deleted", "Event removed.");
+                    await fetchEvents();
+                    setSelectedEvent(null);
+                  } catch (err) {
+                      console.error('Error deleting event:', err);
+                      Alert.alert("Error", "Failed to delete event.");
+                    }
+                }}/>
+            </View>
+          <View style={{ marginTop: 10 }}>
+            <Button title="Close" onPress={() => setSelectedEvent(null)} />
+          </View>
+          </View>
+        </View>
+        </Modal>
+
+
+        </CalendarProvider>
+        <View className='my-10 py-10'>
+          <Button title = "Add Event" onPress={() => setFVisible(true)}></Button>
+        </View>
+      </View>
+    </SafeAreaView>
+  );
+};
+
+export default AgendaInfiniteListScreen;
+
 const styles = StyleSheet.create({
-  containerScroll: {
+  calendar: {
+    paddingLeft: 20,
+    paddingRight: 20
+  },
+  header: {
+    backgroundColor: 'white'
+  },
+  modalContainer: {
     flex: 1,
-    paddingTop: StatusBar.currentHeight,
+    backgroundColor: 'rgba(0,0,0,0.3)',
+    justifyContent: 'center',
+    alignItems: 'center'
   },
-  scrollView: {
+  modalContent: {
     backgroundColor: 'white',
-    marginBottom:100,
+    padding: 20,
+    width: '80%',
+    borderRadius: 8
   },
-  container: {
-    padding: 16,
-  },
-  dayContainer: {
-    marginBottom: 24,
-    borderBottomWidth: 1,
-    paddingBottom: 12,
-  },
-  dayTitle: {
+  modalTitle: {
     fontSize: 18,
     fontWeight: 'bold',
-    marginBottom: 8,
-  },
-  row: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: 8,
+    marginBottom: 10
   },
   input: {
-    flex: 1,
-    borderBottomWidth: 1,
-    padding: 4,
-    marginLeft: 8,
+    borderWidth: 1,
+    borderColor: '#ccc',
+    padding: 8,
+    marginBottom: 10,
+    borderRadius: 4
   },
-  timeBox: {
+  section: {
+    backgroundColor: lightThemeColor,
+    color: 'black',
+    fontSize: 20,
+    textTransform: 'capitalize'
+  },
+  label: {
+    fontWeight: 'bold',
+    marginBottom: 4,
+  },
+  pickerInput: {
+    borderWidth: 1,
+    borderColor: '#ccc',
     padding: 10,
-    backgroundColor: '#ddd',
+    marginBottom: 10,
     borderRadius: 4,
-  },
-  scroll: {
-    padding: 16,
-    flexGrow: 1,
   },
 });
